refactor staking code
